+++ doc/CHANGES.rst
@@ -18,11 +18,16 @@ Future
 
     this version can be uncapable with the previous version
 
+.. note::
+
+    Works only with AnyBlok 0.5.0 and after
+
 * [REF] Add entry point to add new pyramid includem and settings
 * [DEL] properties decorator, it is useless because pyramid have a better
   behaviour
 * [REF] add workingset to define overwritable callback used for application,
   no for the blok, add first callback, get_registry
+* [REF] unit test cause of new version of AnyBlok 0.5.0
 
 0.3.2
 -----